Implement Shopping Cart React models
The G&L frontend now supports a shopping cart, whose state is managed by
session variables on the server. Every user interaction on the cart
(add, remove, update, clear) is sent as a request to the server; once
the client receives the updated cart, the frontend updates all the
corresponding datastores and views. [Finishes #119419677]

Each time the user adds an item to the cart, the app immediately
redirects to the cart page. [Finishes #119419727]

As part of this commit, the API utilities includes a change that adds
the client's credentials in the XHR requests in order to properly set th
session cookie while communicating with the server. Without this change,
every request coming from the frontend will have a different session ID
assigned to it. [Finishes #119783301]